requirements.txt
praw>=7.7.1
pinecone>=5.0.0
google-genai>=0.5.0
pyyaml>=6.0
fastapi>=0.104.0
fastapi-mcp>=0.3.4
uvicorn>=0.24.0
python-multipart>=0.0.6
